Refactor: Improve Scrolling Behavior and Extract Send/Regenerate Functions
- Enhance scrolling behavior by adding a scroll-down button when the user is not at the bottom of the chat.
- Extract `handleSend` and `handleRegenerateResponse` functions into a separate helper file for better organization and reusability.
- Perform necessary cleanup for improved readability and maintenance.
